Add build tags for library/non-library, and arch-specific builds
